Addition of scripts for both flavors. Working dev flavor
Changes to the Vagrantfile to include flavors for dev and release.
Default flavor is now release

Release flavor is still under development and should not be used.

Resizing of the disk size is done by a shell script instead of a Vagrant
script (root-common-resize-disk.sh)

Moved the commands of the root-release-bootstrap to root-dev-bootstrap
as that was the correct place

Changes the script of root-release-bootstrap to receive the .deb file
from github.

Machine reboot is now done by vagrant instead of a reboot command in a
shell script.

Vagrantfile updated to accomodate above changes
Readme file update to describe both flavors.

Changed the git repo flag for new line encoding from CRLF to LF.